+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.1.3
+
+- Improve handling of study session items to ensure that sessions correctly process associated study items, including completing items when a session ends
+- Improve session page UI
+- Add mobile navigation drawer for easier access to navigation on small screens
+- Add work ahead week feature to allow users to complete study sessions in advance
+
 ## v0.1.2
 
 - Improve resource handling in the UI to correctly display associated study items and make assignments